pyproject.toml
[project]
name = "jira-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dagster==1.9.9",
    "dagster-embedded-elt>=0.23.7",
    "dagster-webserver==1.9.9",
    "dlt[duckdb]>=1.5.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
]



